Update index.md
In English language. For Portuguese language, create an issue on this project on GitHub or contact us via email, I relaunched the Gaia project, with an new name. The B2G is the Mozilla community project. I'd forked this project and will be the archived content as in 2014, releasing upon approval of Mozilla to transfer the forked content.